+++ components/Loading.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+import styles from "../styles/loading.module.css";
+
+const Loading: React.FC<{}> = ({}) => {
+	return (
+		<>
+			<div className={styles.ldsGrid}>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+				<div></div>
+			</div>
+		</>
+	);
+};
+
+export default Loading;