+++ scripts/copy-human-assets-to-public.cjs
@@ -19,12 +19,8 @@ function main() {
 			"NEXT_PUBLIC_HUMAN_MODELS_PATH environment variable is not set"
 		);
 	}
-	if (!process.env.NEXT_PUBLIC_TENSOR_WASM_PATH) {
-		throw new Error(
-			"NEXT_PUBLIC_TENSOR_WASM_PATH environment variable is not set"
-		);
-	}
-	const model = {
+
+	const modelBuild = {
 		from: path.resolve(
 			__dirname,
 			"../node_modules/@vladmandic/human-models/models"
@@ -34,18 +30,15 @@ function main() {
 			`../.next/standalone/public/${process.env.NEXT_PUBLIC_HUMAN_MODELS_PATH}`
 		),
 	};
-	const tensorWasm = {
-		from: path.resolve(
-			__dirname,
-			"../node_modules/@tensorflow/tfjs-backend-wasm/dist"
-		),
+
+	const modelDev = Object.assign({}, modelBuild, {
 		to: path.resolve(
 			__dirname,
-			`../.next/standalone/public/${process.env.NEXT_PUBLIC_TENSOR_WASM_PATH}`
+			`../public/${process.env.NEXT_PUBLIC_HUMAN_MODELS_PATH}`
 		),
-	};
+	});
 
-	copyFolderSync(model.from, model.to);
-	copyFolderSync(tensorWasm.from, tensorWasm.to);
+	copyFolderSync(modelBuild.from, modelBuild.to);
+	copyFolderSync(modelDev.from, modelDev.to);
 }
 main();