+++ store/index.ts
@@ -5,7 +5,7 @@ import MathUtils from "../lib/math-utils";
 const HISTORY_SIZE_LIMIT_FRAMES = 20;
 const ROTATION_THRESHOLD_DEGRESS = 0.05;
 const DISTANCE_THRESHOLD_METERS = 0.15;
-const STANDSTILL_THRESHOLD_MS = 2000;
+export const STANDSTILL_THRESHOLD_MS = 2000;
 
 export type EyesOfAIStore = {
 	ready: boolean;
@@ -40,6 +40,12 @@ export type EyesOfAIStore = {
 
 	humanDetected: boolean;
 	setHumanDetected: (humanDetected: boolean) => void;
+
+	humanCloseEnough: boolean;
+	setHumanCloseEnough: (humanCloseEnough: boolean) => void;
+
+	meanDistance: number;
+	setMeanDistance: (meanDistance: number) => void;
 
 	generatedImageExpired: boolean;
 	setGeneratedImageExpired: (generatedImageExpired: boolean) => void;
@@ -86,6 +92,12 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 	setFirstStandStillTime: (firstStillTime) =>
 		set(() => ({ firstStandStillTime: firstStillTime })),
 
+	humanCloseEnough: false,
+	setHumanCloseEnough: (humanCloseEnough) => set(() => ({ humanCloseEnough })),
+
+	meanDistance: 0,
+	setMeanDistance: (meanDistance) => set(() => ({ meanDistance })),
+
 	humanDetected: false,
 	setHumanDetected: (humanDetected) => set(() => ({ humanDetected })),
 
@@ -99,7 +111,6 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 			msInStandStill: 0,
 			firstStandStillTime: undefined,
 			humanDetected: false,
-			human: undefined,
 			generatedImageExpired: false,
 			result: undefined,
 			resultHistory: [],
@@ -108,9 +119,8 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 	trigger: false,
 	checkIfShouldTrigger: () => {
 		const resultHistory = get().resultHistory;
-		const currentResult = get().result;
 
-		if (currentResult.face.length === 0) {
+		if (!hasConsistentlyOneFace(resultHistory)) {
 			set(() => ({
 				humanDetected: false,
 				trigger: false,
@@ -121,10 +131,6 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 		}
 		set(() => ({ humanDetected: true }));
 
-		if (!hasConsistentlyOneFace(currentResult, resultHistory)) {
-			return;
-		}
-
 		const faces = resultHistory
 			.filter((result) => result.face && result.face!.length > 0)
 			.map((result) => result.face![0]);
@@ -132,17 +138,17 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 			return;
 		}
 
-		const distances = faces.map((face) => face.distance);
-		const rolls = faces.map((face) => face.rotation.angle.roll);
-		const pitches = faces.map((face) => face.rotation.angle.pitch);
-		const yaws = faces.map((face) => face.rotation.angle.yaw);
-
-		if (
-			distances.length === 0 ||
-			rolls.length === 0 ||
-			pitches.length === 0 ||
-			yaws.length === 0
-		) {
+		const distances = faces.map((face) => (face.distance ? face.distance : 0));
+		const rolls = faces.map((face) => face.rotation?.angle.roll);
+		const pitches = faces.map((face) => face.rotation?.angle.pitch);
+		const yaws = faces.map((face) => face.rotation?.angle.yaw);
+		const meanDistance = MathUtils.mean(distances);
+		const humanCloseEnough = meanDistance < 0.7;
+		set(() => ({
+			humanCloseEnough: humanCloseEnough,
+			meanDistance: meanDistance,
+		}));
+		if (!humanCloseEnough) {
 			return;
 		}
 
@@ -177,26 +183,9 @@ export const useEyesOfAIStore = create<EyesOfAIStore>()((set, get) => ({
 	},
 }));
 
-function hasConsistentlyOneFace(
-	currentResult: Partial<Result>,
-	resultHistory: Partial<Result>[]
-) {
-	if (currentResult.face.length === 0) {
-		return false;
-	}
-
-	if (!currentResult.face[0]?.age) {
-		return false;
-	}
-
-	if (resultHistory.length === 0) {
-		return false;
-	}
-
-	return (
-		resultHistory.length < HISTORY_SIZE_LIMIT_FRAMES ||
-		resultHistory.every(
-			(result) => result.face.length === currentResult.face.length
-		)
-	);
+function hasConsistentlyOneFace(resultHistory: Partial<Result>[]) {
+	const facesInHistory = resultHistory.filter(
+		(r) => r.face && r.face.length == 1
+	).length;
+	return facesInHistory === resultHistory.length;
 }